+++ R/doiplot.R
@@ -3,10 +3,10 @@
 #' Implementation of the Doi plot proposed by Furuya-Kanamori
 #' et al. (2018) to evaluate bias in meta-analysis.
 #' 
-#' @param TE An object of class \code{lfkindex} or estimated treatment
-#'   effect in individual studies.
+#' @param TE An object of class \code{lfkindex} or \code{meta} or
+#'   estimated treatment effect in individual studies.
 #' @param seTE Standard error of estimated treatment effect (mandatory
-#'   if \code{TE} not of class \code{meta}).
+#'   if \code{TE} not of class \code{lfkindex} or \code{meta}).
 #' @param xlim The x limits (min,max) of the plot.
 #' @param ylim The y limits (min,max) of the plot.
 #' @param xlab A label for the x-axis.
@@ -25,7 +25,6 @@
 #'   \code{\link{funnel.meta}}
 #' 
 #' @references
-#' 
 #' Furuya-Kanamori L, Barendregt JJ, Doi SAR (2018):
 #' A new improved graphical and quantitative method for detecting bias
 #' in meta-analysis.
@@ -79,7 +78,7 @@ doiplot <- function(TE, seTE, xlim, ylim,
        ylim = ylim, xlab = xlab, ylab = ylab,
        ...)
   ##
-  if (lfkindex != "")
+  if (lfkindex)
     legend(pos.lfkindex,
            paste("LFK index", round(lfk$lfkindex, 2)))